* more unsigned int -> size_t changes for safety and performance
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4075 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ extras/collision/collision_dll/vmap/VMapManager.h
@@ -1,5 +1,6 @@
 /* 
  * Copyright (C) 2005,2006,2007 MaNGOS <http://www.mangosproject.org/>
+ * Copyright (C) 2007-2008 Ascent Team <http://www.ascentemu.com/>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -82,7 +83,7 @@ namespace VMAP
 
             // Store all the map tile idents that are loaded for that map
             // some maps are not splitted into tiles and we have to make sure, not removing the map before all tiles are removed
-            Table<unsigned int, bool> iLoadedMapTiles;
+            Table<size_t, bool> iLoadedMapTiles;
             std::string iBasePath;
 
         private:
@@ -90,10 +91,10 @@ namespace VMAP
             bool isAlreadyLoaded(const std::string& pName) { return(iLoadedModelContainer.containsKey(pName)); }
             RayIntersectionIterator<AABSPTree<ModelContainer*>::Node, ModelContainer*> beginRayIntersection(const Ray& ray, bool skipAABoxTests = false) const;
             RayIntersectionIterator<AABSPTree<ModelContainer*>::Node, ModelContainer*> endRayIntersection() const;
-            void setLoadedMapTile(unsigned int pTileIdent) { iLoadedMapTiles.set(pTileIdent, true); }
-            void removeLoadedMapTile(unsigned int pTileIdent) { iLoadedMapTiles.remove(pTileIdent); }
+            void setLoadedMapTile(size_t pTileIdent) { iLoadedMapTiles.set(pTileIdent, true); }
+            void removeLoadedMapTile(size_t pTileIdent) { iLoadedMapTiles.remove(pTileIdent); }
             bool hasLoadedMapTiles() { return(iLoadedMapTiles.size() > 0); }
-            bool containsLoadedMapTile(unsigned int pTileIdent) { return(iLoadedMapTiles.containsKey(pTileIdent)); }
+            bool containsLoadedMapTile(size_t pTileIdent) { return(iLoadedMapTiles.containsKey(pTileIdent)); }
         public:
             ManagedModelContainer *getModelContainer(const std::string& pName) { return(iLoadedModelContainer.get(pName)); }
             const bool hasDirFile(const std::string& pDirName) const { return(iLoadedDirFiles.containsKey(pDirName)); }
@@ -108,9 +109,9 @@ namespace VMAP
 			bool isOutDoors(const Vector3& pos);
             float getHeight(const Vector3& pPos);
 
-            bool loadMap(const std::string& pDirFileName, unsigned int pMapTileIdent);
+            bool loadMap(const std::string& pDirFileName, size_t pMapTileIdent);
             void addModelConatiner(const std::string& pName, ManagedModelContainer *pMc);
-            void unloadMap(const std::string& dirFileName, unsigned int pMapTileIdent);
+            void unloadMap(const std::string& dirFileName, size_t pMapTileIdent);
 
             void getModelContainer(Array<ModelContainer *>& pArray ) { iTree->getMembers(pArray); }
             inline const void addDirFile(const std::string& pDirName, const FilesInDir& pFilesInDir) { iLoadedDirFiles.set(pDirName, pFilesInDir); }
